Allow the ShapefileHarvester to deal with the case where a full zip is provided.  Unfortunately only one Shapefile (with its sidecar files) can be in the zip. Enhanced WPS request document parsing in MainWPSResource to support an external OWS Context file in the proper, standards-compliant way. Lifted parts of ShapefileHarvester and KMLHarvester into a common superclass AbstractFeatureHarvester so that KMLHarvester can more easily access the include/exclude functionality formerly in ShapefileHarvester.

+++ src/net/compusult/geopackage/service/geopackager/GeoPackagingPool.java
@@ -50,7 +50,9 @@ public class GeoPackagingPool implements InitializingBean {
 		 * get rid of them.
 		 */
 		for (File existingFile : new File(workDirectory).listFiles()) {
-			existingFile.delete();
+			if (existingFile.isFile()) {
+				existingFile.delete();
+			}
 		}
 	}
